requirements.txt
pytest==5.0.1
allure-pytest==2.7.1
pytest-ordering==0.6
ifacetools==0.0.9
requests==2.22.0
redis==3.3.5
Faker==2.0.0
unicode==2.7
factory-boy==2.12.0
sqlalchemy==1.3.7
mysql-connector-python==8.0.17
